+++ caddy-s3-nomad/caddy-s3-nomad.d/local/share/cook/bin/configure-acme.sh
@@ -9,11 +9,10 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 SCRIPT=$(readlink -f "$0")
 TEMPLATEPATH=$(dirname "$SCRIPT")/../templates
-
 
 # make directories if they don't exist
 mkdir -p /mnt/acme