+++ caddy-s3-nomad/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Version bump for new base image
 * Update to 2025Q1 for pkg sources
 * Update for rsync security issue
+* Fix switch to latest over quarterly
 
 ---