+++ caddy-s3-nomad/CHANGELOG.md
@@ -1,6 +1,7 @@
 0.18
 
 * Version bump for new base image
+* Add certificate expiry scripts for alertmanager notices
 
 ---