+++ caddy-s3-nomad/CHANGELOG.md
@@ -1,3 +1,10 @@
+0.18
+
+* Version bump for new base image
+* Add certificate expiry scripts for alertmanager notices
+
+---
+
 0.17
 
 * Version bump for new base image